+++ test/processor_test.rb
@@ -39,7 +39,7 @@ class ProcessorTest < Minitest::Test
 
       mock_capture3 = Minitest::Mock.new
       mock_capture3.expect :call, ["", "", mock_status], [
-        { "RUBYOPT" => nil, "GIT_SSH" => (workspace.working_dir / "id_rsa").to_s },
+        { "RUBYOPT" => nil, "GIT_SSH_COMMAND" => "ssh -F '#{workspace.working_dir / 'id_rsa'}'" },
         "ls",
         chdir: workspace.working_dir.to_s, stdin_data: nil
       ]
@@ -387,14 +387,14 @@ class ProcessorTest < Minitest::Test
     with_workspace do |workspace|
       processor = new_processor(workspace: workspace)
 
-      assert_equal({ "RUBYOPT" => nil, "GIT_SSH" => nil }, processor.env_hash)
+      assert_equal({ "RUBYOPT" => nil, "GIT_SSH_COMMAND" => nil }, processor.env_hash)
 
       processor.push_env_hash({ "RBENV_VERSION" => "2.5.0" }) do
-        assert_equal({ "RUBYOPT" => nil, "GIT_SSH" => nil, "RBENV_VERSION" => "2.5.0" },
+        assert_equal({ "RUBYOPT" => nil, "GIT_SSH_COMMAND" => nil, "RBENV_VERSION" => "2.5.0" },
                      processor.env_hash)
       end
 
-      assert_equal({ "RUBYOPT" => nil, "GIT_SSH" => nil }, processor.env_hash)
+      assert_equal({ "RUBYOPT" => nil, "GIT_SSH_COMMAND" => nil }, processor.env_hash)
     end
   end